Move away Cors enum
